separated SubstanceDetail component
