+++ frontend/src/components/ChemSubTable.tsx
@@ -83,141 +83,6 @@ const SubstanceCard = ({units, substance, selected, setSelected, setDetail}) => 
     )
 }
 
-const SubstanceDetail = ({units, substance, setDetail}) => {
-    // let [edit, setEdit] = useState(false)
-    if (substance) return(
-        <Modal
-          show={substance !== null}
-          onHide={() => setDetail(null)}
-          dialogClassName="modal-90w"
-          centered
-        >
-          <Modal.Header closeButton>
-            <Modal.Title>
-                {substance.name}
-            </Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Container>
-              <Row>
-                {substance.image
-                    ? <Col xs={12} lg={2} className={"d-flex justify-content-center"}><SubstanceImage path={substance.image}/></Col>
-                    : null
-                }
-                <Col xs={12} lg={10}>
-                    <Table borderless>
-                        <tbody>
-                            <DataRow
-                                name={"formula"}
-                                data={substance.formula}
-                                text={<FormulaFormatter formula={substance.formula}/>}
-                            />
-                            <DataRow
-                                name={"CAS"}
-                                data={substance.cas}
-                            />
-                            <DataRow
-                                name={"PubChem SID"}
-                                data={substance.pubchem_sid}
-                            />
-                            <DataRow
-                                name={"PubChem CID"}
-                                data={substance.pubchem_cid}
-                            />
-                            <DataRow
-                                name={"InChI"}
-                                data={substance.inchi}
-                            />
-                            <DataRow
-                                name={"InChI key"}
-                                data={substance.inchi_key}
-                            />
-                            <DataRow
-                                name={"SMILES"}
-                                data={substance.smiles}
-                            />
-                        </tbody>
-                    </Table>
-                </Col>
-              </Row>
-            </Container>
-            <Tabs
-              defaultActiveKey="home"
-              transition={false}
-              className="mb-3"
-            >
-              <Tab eventKey="home" title="Properties">
-                <Container>
-                  <Row>
-                    <Col xs={12} lg={10}>
-                        <Table borderless>
-                            <tbody>
-                                <DataRow
-                                    name={"molar mass"}
-                                    data={substance.mol_weight}
-                                    text={`${substance.mol_weight} ${
-                                        units.find(u => u.id === substance.mol_weight_unit).symbol
-                                    }`}
-                                />
-                                <DataRow
-                                    name={"exact mass"}
-                                    data={substance.exact_mass}
-                                    text={`${substance.exact_mass} ${
-                                        units.find(u => u.id === substance.exact_mass_unit).symbol
-                                    }`}
-                                />
-                            </tbody>
-                        </Table>
-                    </Col>
-                  </Row>
-                </Container>
-              </Tab>
-              <Tab eventKey="profile" title="Spectra">
-                  <p>
-                    Ipsum molestiae natus adipisci modi eligendi? Debitis amet quae unde
-                    commodi aspernatur enim, consectetur. Cumque deleniti temporibus
-                    ipsam atque a dolores quisquam quisquam adipisci possimus
-                    laboriosam. Quibusdam facilis doloribus debitis! Sit quasi quod
-                    accusamus eos quod. Ab quos consequuntur eaque quo rem! Mollitia
-                    reiciendis porro quo magni incidunt dolore amet atque facilis ipsum
-                    deleniti rem!
-                  </p>
-              </Tab>
-              <Tab eventKey="contact" title="SDS">
-                  <p>
-                    Ipsum molestiae natus adipisci modi eligendi? Debitis amet quae unde
-                    commodi aspernatur enim, consectetur. Cumque deleniti temporibus
-                    ipsam atque a dolores quisquam quisquam adipisci possimus
-                    laboriosam. Quibusdam facilis doloribus debitis! Sit quasi quod
-                    accusamus eos quod. Ab quos consequuntur eaque quo rem! Mollitia
-                    reiciendis porro quo magni incidunt dolore amet atque facilis ipsum
-                    deleniti rem!
-                  </p>
-              </Tab>
-            </Tabs>
-            {/*<Form>*/}
-            {/*  <Form.Group className="mb-3" controlId="exampleForm.ControlInput1">*/}
-            {/*    <Form.Label>Email address</Form.Label>*/}
-            {/*    <Form.Control*/}
-            {/*      type="email"*/}
-            {/*      name="email2"*/}
-            {/*      placeholder="name@example.com"*/}
-            {/*      autoFocus*/}
-            {/*    />*/}
-            {/*  </Form.Group>*/}
-            {/*  <Form.Group*/}
-            {/*      className="mb-3"*/}
-            {/*      controlId="exampleForm.ControlTextarea1"*/}
-            {/*  >*/}
-            {/*    <Form.Label>Example textarea</Form.Label>*/}
-            {/*    <Form.Control as="textarea" rows={3} />*/}
-            {/*  </Form.Group>*/}
-            {/*</Form>*/}
-          </Modal.Body>
-        </Modal>
-    )
-}
-
 const ChemSubTable = ({units, substances, selected, setSelected}) => {
     let [detail, setDetail] = useState<Substance | null>(null)
     if (substances.length === 0) return(<Row>No data available.</Row>)
